Modify load order of locales
The load order as controlled by an additional code snippet during
Redmine initialization is modified to be conditional. That is, the
code won't get executed when the redmine_message_customize plugin is not
installed anymore.

This won't break the restart when the plugin is (temporarily)
uninstalled but the additional file in Redmine config/initializers
is not removed.